Add files via upload
Two numbers are taken from the user as input in this promise. If both the numbers are even the promise is fulfilled and resolves to success otherwise it is rejected and displays error. I have used the async- await method to resolve the promise and try and catch method is used to handle the result of the promise.