async_await.js
I have used the async-await method to resolve the promise created in this code. The promise is such that the user is asked to enter two numbers and if both of them are even then the promise is fulfilled and resolves to success otherwise it is rejected and displays error. I have used the try and catch method to handle the result of the promise.